test/backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
paddlepaddle
paddleocr
pillow
reportlab
requests
numpy
pydantic
python-dotenv
